+++ src/pages/Home/components/feed/PostDetailItem.tsx
@@ -1,0 +1,9 @@
+const PostDetailItem = () => {
+
+    return (
+        <div>
+        </div>
+    )
+}
+
+export default PostDetailItem;
