+++ src/pages/Home/components/postDetail/index.tsx
@@ -1,0 +1,26 @@
+import { getPostDetail } from '@/apis/postApis';
+import { useState,useEffect } from 'react';
+
+type PostDetail = {
+    id : number,
+}
+
+const PostDetail = ({id} : PostDetail) => {
+
+    const [postDetails, setPostDetails] =useState(null);
+    useEffect(()=>{
+        (async()=>{
+            const details = await getPostDetail(id);
+            setPostDetails(details);
+        })();
+
+    },[]);
+
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default PostDetail;
